Addressed code review
