+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/HiveTypeCoercion.scala
@@ -79,6 +79,7 @@ trait HiveTypeCoercion {
     CaseWhenCoercion ::
     Division ::
     PropagateTypes ::
+    ExpectedInputConversion ::
     Nil
 
   /**
@@ -640,6 +641,31 @@ trait HiveTypeCoercion {
           // Types match up.  Hopefully some other rule fixes whatever is wrong with resolution.
           cw
         }
+    }
+  }
+
+  /**
+   * Casts types according to the expected input types for Expressions that have the trait
+   * `ExpectsInputTypes`.
+   */
+  object ExpectedInputConversion extends Rule[LogicalPlan] {
+
+    def apply(plan: LogicalPlan): LogicalPlan = plan transform {
+      case q: LogicalPlan => q transformExpressions {
+        // Skip nodes who's children have not been resolved yet.
+        case e if !e.childrenResolved => e
+
+        case e: ExpectsInputTypes if e.children.map(_.dataType) != e.expectedChildTypes =>
+          val newC = (e.children, e.children.map(_.dataType), e.expectedChildTypes).zipped.map {
+            case (child, actual, expected) =>
+              if (actual == expected) {
+                child
+              } else {
+                Cast(child, expected)
+              }
+          }
+          e.withNewChildren(newC)
+      }
     }
   }
