+++ sql/core/src/test/scala/org/apache/spark/sql/ColumnExpressionSuite.scala
@@ -18,7 +18,6 @@
 package org.apache.spark.sql
 
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.mathfunctions._
 import org.apache.spark.sql.test.TestSQLContext
 import org.apache.spark.sql.test.TestSQLContext.implicits._
 import org.apache.spark.sql.types._
@@ -330,187 +329,5 @@ class ColumnExpressionSuite extends QueryTest {
       .schema
     assert(schema("value").metadata === Metadata.empty)
     assert(schema("abc").metadata === metadata)
-  }
-
-  def testOneToOneMathFunction[@specialized(Int, Double, Float, Long) T]
-      (c: Column => Column, f: T => T): Unit = {
-    checkAnswer(
-      doubleData.select(c('a)).orderBy('a.asc),
-      (1 to 100).map(n => Row(f((n * 0.02 - 1).asInstanceOf[T])))
-    )
-
-    checkAnswer(
-      doubleData.select(c('b)).orderBy('b.desc),
-      (1 to 100).map(n => Row(f((-n * 0.02 + 1).asInstanceOf[T])))
-    )
-
-    checkAnswer(
-      doubleData.select(c(lit(null))),
-      (1 to 100).map(_ => Row(null))
-    )
-  }
-
-  test("sin") {
-    testOneToOneMathFunction(sin, math.sin)
-  }
-
-  test("asin") {
-    testOneToOneMathFunction(asin, math.asin)
-  }
-
-  test("sinh") {
-    testOneToOneMathFunction(sinh, math.sinh)
-  }
-
-  test("cos") {
-    testOneToOneMathFunction(cos, math.cos)
-  }
-
-  test("acos") {
-    testOneToOneMathFunction(acos, math.acos)
-  }
-
-  test("cosh") {
-    testOneToOneMathFunction(cosh, math.cosh)
-  }
-
-  test("tan") {
-    testOneToOneMathFunction(tan, math.tan)
-  }
-
-  test("atan") {
-    testOneToOneMathFunction(atan, math.atan)
-  }
-
-  test("tanh") {
-    testOneToOneMathFunction(tanh, math.tanh)
-  }
-
-  test("toDeg") {
-    testOneToOneMathFunction(toDeg, math.toDegrees)
-  }
-
-  test("toRad") {
-    testOneToOneMathFunction(toRad, math.toRadians)
-  }
-
-  test("cbrt") {
-    testOneToOneMathFunction(cbrt, math.cbrt)
-  }
-
-  test("ceil") {
-    testOneToOneMathFunction(ceil, math.ceil)
-  }
-
-  test("floor") {
-    testOneToOneMathFunction(floor, math.floor)
-  }
-
-  test("rint") {
-    testOneToOneMathFunction(rint, math.rint)
-  }
-
-  test("exp") {
-    testOneToOneMathFunction(exp, math.exp)
-  }
-
-  test("expm1") {
-    testOneToOneMathFunction(expm1, math.expm1)
-  }
-
-  test("signum") {
-    testOneToOneMathFunction[Double](signum, math.signum)
-  }
-
-  test("isignum") {
-    testOneToOneMathFunction[Int](isignum, math.signum)
-  }
-
-  test("fsignum") {
-    testOneToOneMathFunction[Float](fsignum, math.signum)
-  }
-
-  test("lsignum") {
-    testOneToOneMathFunction[Long](lsignum, math.signum)
-  }
-
-  def testTwoToOneMathFunction(
-      c: (Column, Column) => Column,
-      d: (Column, Double) => Column,
-      f: (Double, Double) => Double): Unit = {
-    checkAnswer(
-      nnDoubleData.select(c('a, 'a)).orderBy('a.asc),
-      nnDoubleData.collect().toSeq.map(r => Row(f(r.getDouble(0), r.getDouble(0))))
-    )
-
-    checkAnswer(
-      nnDoubleData.select(c('a, 'b)).orderBy('a.asc),
-      nnDoubleData.collect().toSeq.map(r => Row(f(r.getDouble(0), r.getDouble(1))))
-    )
-
-    checkAnswer(
-      nnDoubleData.select(d('a, 2.0)).orderBy('a.asc),
-      nnDoubleData.collect().toSeq.map(r => Row(f(r.getDouble(0), 2.0)))
-    )
-
-    checkAnswer(
-      nnDoubleData.select(d('a, -0.5)).orderBy('a.asc),
-      nnDoubleData.collect().toSeq.map(r => Row(f(r.getDouble(0), -0.5)))
-    )
-
-    val nonNull = nullDoubles.collect().toSeq.filter(r => r.get(0) != null)
-
-    checkAnswer(
-      nullDoubles.select(c('a, 'a)).orderBy('a.asc),
-      Row(null) +: nonNull.map(r => Row(f(r.getDouble(0), r.getDouble(0))))
-    )
-  }
-
-  test("pow") {
-    testTwoToOneMathFunction(pow, pow, math.pow)
-  }
-
-  test("hypot") {
-    testTwoToOneMathFunction(hypot, hypot, math.hypot)
-  }
-
-  test("atan2") {
-    testTwoToOneMathFunction(atan2, atan2, math.atan2)
-  }
-
-  def testOneToOneNonNegativeMathFunction(c: Column => Column, f: Double => Double): Unit = {
-    checkAnswer(
-      testData.select(c('key)).orderBy('key.asc),
-      (1 to 100).map(n => Row(f(n)))
-    )
-
-    if (f(-1) === math.log1p(-1)) {
-      checkAnswer(
-        negativeData.select(c('key)).orderBy('key.desc),
-        Row(Double.NegativeInfinity) +: (2 to 100).map(n => Row(null))
-      )
-    } else {
-      checkAnswer(
-        negativeData.select(c('key)).orderBy('key.desc),
-        (1 to 100).map(n => Row(null))
-      )
-    }
-
-    checkAnswer(
-      testData.select(c(lit(null))),
-      (1 to 100).map(_ => Row(null))
-    )
-  }
-
-  test("log") {
-    testOneToOneNonNegativeMathFunction(log, math.log)
-  }
-
-  test("log10") {
-    testOneToOneNonNegativeMathFunction(log10, math.log10)
-  }
-
-  test("log1p") {
-    testOneToOneNonNegativeMathFunction(log1p, math.log1p)
   }
 }
